+++ lib/lineworks/api/event/join.rb
@@ -1,0 +1,11 @@
+module Lineworks
+  module Api
+    module Event
+      # Event object for when your LINE WORKS official account joins a group or room.
+      #
+      # @see: https://developers.worksmobile.com/jp/docs/bot-callback-join
+      class Join < Base
+      end
+    end
+  end
+end