+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+## [?.?.?] - 2024-01-02
+
+- Add Event class and its subclasses
+
 ## [0.1.0] - 2024-01-02
 
 - Initial release